Sửa lỗi và thêm ràng buộc cho các tính năng thêm, xóa, sửa

+++ QuanLyNhanSuApp_X/FormCheDo_BaoHiem.cs
@@ -216,9 +216,25 @@ namespace QuanLyNhanSuApp
                 message += "Thời gian hiệu lực của bảo hiểm không hợp lệ!\n";
                 
             }
-            else
+            if (txtMaNV.Text.Length > 10)
             {
-                message += "Ngày kết thúc phải lớn hơn ngày cấp phát!\n";
+                message += "Mã nhân viên chỉ chứa tối đa 10 ký tự\n";
+            }
+            if (txtLoaiBaoHiem.Text.Length > 50)
+            {
+                message += "Loại bảo hiểm quá dài\n";
+            }
+            if (txtGhiChu.Text.Length > 50)
+            {
+                message += "Ghi chú quá dài\n";
+            }
+            if (txtMaBaoHiem.Text.Length > 10)
+            {
+                message += "Mã bảo hiểm chỉ chứa tối đa 10 ký tự\n";
+            }
+            if (txtMaSoBaoHiem.Text.Length > 20)
+            {
+                message += "Mã số bảo hiểm quá dài\n";
             }
             return message;
         }
@@ -233,25 +249,34 @@ namespace QuanLyNhanSuApp
             {
                 if (isValid() == "")
                 {
-                    try
+                    DataAccess dataAccess = new DataAccess();
+                    DataTable dataTable = dataAccess.GetData("SELECT maBaoHiem FROM baohiem WHERE maBaoHiem = '" + txtMaBaoHiem.Text + "'");
+                    if (dataTable.Rows.Count != 0)
                     {
-                        string timeInString = dateTimePickerNgayCap.Value.ToString("yyyy-MM-dd");
-
-                        string timeOutString = dateTimePickerNgayKetThuc.Value.ToString("yyyy-MM-dd");
-
-                        DataAccess dataAccess = new DataAccess();
-                        string query = "INSERT INTO `employeems`.`baohiem` (`maNV`, `loaiBaoHiem`, `maBaoHiem`, `ngayCap`, `noiCap`, `ghiChu`, `ngayKetThuc`, `maSoBH`) " +
-                            "VALUES ('"+txtMaNV.Text+"', '"+txtLoaiBaoHiem.Text+"', '"+txtMaBaoHiem.Text+"', '"+timeInString+"', '"+txtNoiCap.Text+"', '"+txtGhiChu.Text+"', '"+timeOutString+"', '"+txtMaSoBaoHiem.Text+"');";
-                        dataAccess.InsertData(query);
-                        MessageBox.Show("Thêm thành công!", "Thông báo");
-                        BindToDataGridView();
-
-
+                        MessageBox.Show("Mã bảo hiểm đã tồn tại!", "Cảnh báo");
                     }
-                    catch (Exception ex)
+                    else
                     {
-                        MessageBox.Show("Lỗi: " + ex, "Thông báo");
+                        try
+                        {
+                            string timeInString = dateTimePickerNgayCap.Value.ToString("yyyy-MM-dd");
 
+                            string timeOutString = dateTimePickerNgayKetThuc.Value.ToString("yyyy-MM-dd");
+
+                            dataAccess = new DataAccess();
+                            string query = "INSERT INTO `employeems`.`baohiem` (`maNV`, `loaiBaoHiem`, `maBaoHiem`, `ngayCap`, `noiCap`, `ghiChu`, `ngayKetThuc`, `maSoBH`) " +
+                                "VALUES ('" + txtMaNV.Text + "', '" + txtLoaiBaoHiem.Text + "', '" + txtMaBaoHiem.Text + "', '" + timeInString + "', '" + txtNoiCap.Text + "', '" + txtGhiChu.Text + "', '" + timeOutString + "', '" + txtMaSoBaoHiem.Text + "');";
+                            dataAccess.InsertData(query);
+                            MessageBox.Show("Thêm thành công!", "Thông báo");
+                            BindToDataGridView();
+
+
+                        }
+                        catch (Exception ex)
+                        {
+                            MessageBox.Show("Lỗi: " + ex, "Thông báo");
+
+                        }
                     }
                 }
                 else
@@ -270,21 +295,35 @@ namespace QuanLyNhanSuApp
             }
             else
             {
-                try
+                DataAccess dataAccess = new DataAccess();
+                DataTable dataTable = dataAccess.GetData("SELECT maBaoHiem FROM baohiem WHERE maBaoHiem = '" + txtMaBaoHiem.Text + "'");
+                if (dataTable.Rows.Count == 0)
                 {
-                    txtMaBaoHiem.BackColor = SystemColors.Window;
-                    DataAccess dataAccess = new DataAccess();
-                    string query = "DELETE FROM `employeems`.`baohiem` WHERE (`maBaoHiem` = '" + txtMaBaoHiem.Text + "');";
-                    dataAccess.DeleteData(query);
-                    MessageBox.Show("Xóa thành công", "Thông báo");
-                    BindToDataGridView();
-                    ResetTextBox();
+                    MessageBox.Show("Mã bảo hiểm không tồn tại!", "Cảnh báo");
+                }
+                else
+                {
+                    DialogResult result = MessageBox.Show("Bạn có chắc chắn muốn xóa không?", "Xác nhận xóa", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
 
-                }
-                catch (Exception ex)
-                {
-                    MessageBox.Show("Lỗi: " + ex, "Thông báo");
-                }
+                    if (result == DialogResult.Yes)
+                    {
+                        try
+                        {
+                            txtMaBaoHiem.BackColor = SystemColors.Window;
+                            dataAccess = new DataAccess();
+                            string query = "DELETE FROM `employeems`.`baohiem` WHERE (`maBaoHiem` = '" + txtMaBaoHiem.Text + "');";
+                            dataAccess.DeleteData(query);
+                            MessageBox.Show("Xóa thành công", "Thông báo");
+                            BindToDataGridView();
+                            ResetTextBox();
+
+                        }
+                        catch (Exception ex)
+                        {
+                            MessageBox.Show("Lỗi: " + ex, "Thông báo");
+                        }
+                    }
+                }      
             }
         }
 
@@ -298,34 +337,48 @@ namespace QuanLyNhanSuApp
             {
                 if (isValid() == "")
                 {
-                    try
+                    DataAccess dataAccess = new DataAccess();
+                    DataTable dataTable = dataAccess.GetData("SELECT maBaoHiem FROM baohiem WHERE maBaoHiem = '" + txtMaBaoHiem.Text + "'");
+                    if (dataTable.Rows.Count == 0)
                     {
-
-
-                        string timeInString = dateTimePickerNgayCap.Value.ToString("yyyy-MM-dd");
-
-                        string timeOutString = dateTimePickerNgayKetThuc.Value.ToString("yyyy-MM-dd");
-
-                        DataAccess dataAccess = new DataAccess();
-                        string query = "UPDATE `employeems`.`baohiem` " +
-                            "SET `maNV` = '"+txtMaNV.Text+"', " +
-                            "`loaiBaoHiem` = '"+txtLoaiBaoHiem.Text+"', " +
-                            "`maBaoHiem` = '"+txtMaBaoHiem.Text+"', " +
-                            "`ngayCap` = '"+timeInString+"', " +
-                            "`noiCap` = '"+txtNoiCap.Text+"', " +
-                            "`ghiChu` = '"+txtGhiChu.Text+"', " +
-                            "`ngayKetThuc` = '"+timeOutString+"', " +
-                            "`maSoBH` = '"+txtMaSoBaoHiem.Text+"' WHERE (`maBaoHiem` = '"+txtMaBaoHiem.Text+"');";
-                        dataAccess.UpdateData(query);
-                        MessageBox.Show("Cập nhật thành công!", "Thông báo");
-                        BindToDataGridView();
-
-
+                        MessageBox.Show("Mã bảo hiểm không tồn tại!", "Cảnh báo");
                     }
-                    catch (Exception ex)
+                    else
                     {
-                        MessageBox.Show("Lỗi: " + ex, "Thông báo");
+                        DialogResult result = MessageBox.Show("Bạn có chắc chắn muốn cập nhật không?", "Xác nhận cập nhật", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
 
+                        if (result == DialogResult.Yes)
+                        {
+                            try
+                            {
+
+
+                                string timeInString = dateTimePickerNgayCap.Value.ToString("yyyy-MM-dd");
+
+                                string timeOutString = dateTimePickerNgayKetThuc.Value.ToString("yyyy-MM-dd");
+
+                                dataAccess = new DataAccess();
+                                string query = "UPDATE `employeems`.`baohiem` " +
+                                    "SET `maNV` = '" + txtMaNV.Text + "', " +
+                                    "`loaiBaoHiem` = '" + txtLoaiBaoHiem.Text + "', " +
+                                    "`maBaoHiem` = '" + txtMaBaoHiem.Text + "', " +
+                                    "`ngayCap` = '" + timeInString + "', " +
+                                    "`noiCap` = '" + txtNoiCap.Text + "', " +
+                                    "`ghiChu` = '" + txtGhiChu.Text + "', " +
+                                    "`ngayKetThuc` = '" + timeOutString + "', " +
+                                    "`maSoBH` = '" + txtMaSoBaoHiem.Text + "' WHERE (`maBaoHiem` = '" + txtMaBaoHiem.Text + "');";
+                                dataAccess.UpdateData(query);
+                                MessageBox.Show("Cập nhật thành công!", "Thông báo");
+                                BindToDataGridView();
+
+
+                            }
+                            catch (Exception ex)
+                            {
+                                MessageBox.Show("Lỗi: " + ex, "Thông báo");
+
+                            }
+                        }
                     }
                 }
                 else
@@ -376,6 +429,30 @@ namespace QuanLyNhanSuApp
             FormSearching fsearch = new FormSearching(query, 5,6);
             fsearch.FormClosed += FormSearchClosed;
             fsearch.ShowDialog();
+        }
+
+        private void txtMaNV_KeyPress(object sender, KeyPressEventArgs e)
+        {
+            if (!char.IsLetterOrDigit(e.KeyChar) && !char.IsControl(e.KeyChar))
+            {
+                e.Handled = true; // Không cho phép ký tự được nhập vào
+            }
+        }
+
+        private void txtMaBaoHiem_KeyPress(object sender, KeyPressEventArgs e)
+        {
+            if (!char.IsLetterOrDigit(e.KeyChar) && !char.IsControl(e.KeyChar))
+            {
+                e.Handled = true; // Không cho phép ký tự được nhập vào
+            }
+        }
+
+        private void txtMaSoBaoHiem_KeyPress(object sender, KeyPressEventArgs e)
+        {
+            if (!char.IsLetterOrDigit(e.KeyChar) && !char.IsControl(e.KeyChar))
+            {
+                e.Handled = true; // Không cho phép ký tự được nhập vào
+            }
         }
     }
 }
